requirements.txt
Pillow
hexdump